+++ core-ng/src/main/java/core/framework/internal/cache/CacheManager.java
@@ -14,18 +14,16 @@ import java.util.Optional;
  * @author neo
  */
 public class CacheManager {
-    private final CacheStore cacheStore;
     private final Map<String, CacheImpl<?>> caches = Maps.newHashMap();
 
-    public CacheManager(CacheStore cacheStore) {
-        this.cacheStore = cacheStore;
-    }
+    public CacheStore localCacheStore;
+    public CacheStore remoteCacheStore;
 
-    public <T> Cache<T> add(Class<T> cacheClass, Duration duration) {
+    public <T> Cache<T> add(Class<T> cacheClass, Duration duration, boolean localCache) {
         new CacheClassValidator(cacheClass).validate();
 
         String name = cacheName(cacheClass);
-        CacheImpl<T> cache = new CacheImpl<>(name, cacheClass, duration, cacheStore);
+        CacheImpl<T> cache = new CacheImpl<>(name, cacheClass, duration, localCache ? localCacheStore : remoteCacheStore);
         CacheImpl<?> previous = caches.putIfAbsent(name, cache);
         if (previous != null) throw new Error("found duplicate cache name, name=" + name);
         return cache;