+++ core-ng/src/main/java/core/framework/module/CacheConfig.java
@@ -3,7 +3,6 @@ package core.framework.module;
 import core.framework.cache.Cache;
 import core.framework.http.HTTPMethod;
 import core.framework.internal.cache.CacheManager;
-import core.framework.internal.cache.CacheStore;
 import core.framework.internal.cache.LocalCacheStore;
 import core.framework.internal.cache.RedisCacheStore;
 import core.framework.internal.module.Config;
@@ -26,51 +25,13 @@ public class CacheConfig extends Config {
     private final Logger logger = LoggerFactory.getLogger(CacheConfig.class);
     private ModuleContext context;
     private CacheManager cacheManager;
+    private LocalCacheStore localCacheStore;
 
     @Override
     protected void initialize(ModuleContext context, String name) {
         this.context = context;
-    }
 
-    @Override
-    protected void validate() {
-        if (cacheManager == null || cacheManager.caches().isEmpty()) {
-            throw new Error("cache is configured but no cache added, please remove unnecessary config");
-        }
-    }
-
-    public void local() {
-        if (cacheManager != null)
-            throw new Error("cache is already configured, please configure cache store only once");
-
-        logger.info("create local cache store");
-        LocalCacheStore cacheStore = new LocalCacheStore();
-        context.backgroundTask().scheduleWithFixedDelay(cacheStore::cleanup, Duration.ofMinutes(30));
-
-        configureCacheManager(cacheStore);
-    }
-
-    public void redis(String host) {
-        if (cacheManager != null)
-            throw new Error("cache is already configured, please configure cache store only once");
-
-        configureRedis(host);
-    }
-
-    void configureRedis(String host) {
-        logger.info("create redis cache manager, host={}", host);
-
-        RedisImpl redis = new RedisImpl("redis-cache");
-        redis.host = host;
-        redis.timeout(Duration.ofSeconds(1));   // for cache, use shorter timeout than default redis config
-        context.shutdownHook.add(ShutdownHook.STAGE_7, timeout -> redis.close());
-        context.backgroundTask().scheduleWithFixedDelay(redis.pool::refresh, Duration.ofMinutes(5));
-        context.stat.metrics.add(new PoolMetrics(redis.pool));
-        configureCacheManager(new RedisCacheStore(redis));
-    }
-
-    private void configureCacheManager(CacheStore cacheStore) {
-        cacheManager = new CacheManager(cacheStore);
+        cacheManager = new CacheManager();
 
         var controller = new CacheController(cacheManager);
         context.route(HTTPMethod.GET, "/_sys/cache", (LambdaController) controller::list, true);
@@ -79,11 +40,68 @@ public class CacheConfig extends Config {
         context.route(HTTPMethod.DELETE, "/_sys/cache/:name/:key", (LambdaController) controller::delete, true);
     }
 
-    public <T> void add(Class<T> cacheClass, Duration duration) {
-        if (cacheManager == null) throw new Error("cache is not configured, please configure cache store first");
+    @Override
+    protected void validate() {
+        if (cacheManager.caches().isEmpty()) {
+            throw new Error("cache is configured but no cache added, please remove unnecessary config");
+        }
+    }
 
-        logger.info("add cache, class={}", cacheClass.getCanonicalName());
-        Cache<T> cache = cacheManager.add(cacheClass, duration);
+    public void redis(String host) {
+        if (cacheManager.remoteCacheStore != null)
+            throw new Error("cache is already configured, please configure cache store only once");
+
+        configureRedis(host);
+    }
+
+    public void local() {
+        if (cacheManager.remoteCacheStore != null)
+            throw new Error("cache store is already configured, please configure only once");
+
+        cacheManager.remoteCacheStore = localCacheStore();
+    }
+
+    public void maxLocalSize(long size) {
+        localCacheStore().maxSize = size;
+    }
+
+    public <T> void local(Class<T> cacheClass, Duration duration) {
+        if (cacheManager.localCacheStore == null) {
+            cacheManager.localCacheStore = localCacheStore();
+        }
+
+        logger.info("add local cache, class={}", cacheClass.getCanonicalName());
+        Cache<T> cache = cacheManager.add(cacheClass, duration, true);
         context.beanFactory.bind(Types.generic(Cache.class, cacheClass), null, cache);
+    }
+
+    public <T> void remote(Class<T> cacheClass, Duration duration) {
+        if (cacheManager.remoteCacheStore == null) throw new Error("cache store is not configured, please configure first");
+
+        logger.info("add remote cache, class={}", cacheClass.getCanonicalName());
+        Cache<T> cache = cacheManager.add(cacheClass, duration, false);
+        context.beanFactory.bind(Types.generic(Cache.class, cacheClass), null, cache);
+    }
+
+    void configureRedis(String host) {
+        logger.info("create redis cache store, host={}", host);
+
+        RedisImpl redis = new RedisImpl("redis-cache");
+        redis.host = host;
+        redis.timeout(Duration.ofSeconds(1));   // for cache, use shorter timeout than default redis config
+        context.shutdownHook.add(ShutdownHook.STAGE_7, timeout -> redis.close());
+        context.backgroundTask().scheduleWithFixedDelay(redis.pool::refresh, Duration.ofMinutes(5));
+        context.stat.metrics.add(new PoolMetrics(redis.pool));
+        cacheManager.remoteCacheStore = new RedisCacheStore(redis);
+    }
+
+    private LocalCacheStore localCacheStore() {
+        if (localCacheStore == null) {
+            logger.info("create local cache store");
+            var cacheStore = new LocalCacheStore();
+            context.backgroundTask().scheduleWithFixedDelay(cacheStore::cleanup, Duration.ofMinutes(5));
+            localCacheStore = cacheStore;
+        }
+        return localCacheStore;
     }
 }